bpf: split local-delivery logic from l3.h
Local-delivery and L3-forwarding (== TTL adjustment) are very different
things, let the callers select which parts they need.

In particular the local-delivery parts make no sense from XDP context, so
all the shared code in eg. nodeport.h really shouldn't pull them in via
the l3.h include.

Update the include statements accordingly.
