+++ bpf/bpf_host.c
@@ -37,8 +37,8 @@
 #include "lib/policy.h"
 #include "lib/trace.h"
 #include "lib/identity.h"
-#include "lib/l3.h"
 #include "lib/l4.h"
+#include "lib/local_delivery.h"
 #include "lib/drop.h"
 #include "lib/encap.h"
 #include "lib/nat.h"
