+++ bpf/lib/nodeport.h
@@ -10,6 +10,7 @@
 #include "tailcall.h"
 #include "nat.h"
 #include "edt.h"
+#include "l3.h"
 #include "lb.h"
 #include "common.h"
 #include "overloadable.h"
