+++ bpf/lib/source_info.h
@@ -51,7 +51,7 @@ __id_for_file(const char *const header_name)
 	_strcase_(107, "mcast.h");
 	_strcase_(108, "ipv4.h");
 	_strcase_(109, "conntrack.h");
-	_strcase_(110, "l3.h");
+	_strcase_(110, "local_delivery.h");
 	_strcase_(111, "trace.h");
 	_strcase_(112, "encap.h");
 	_strcase_(113, "encrypt.h");
